+++ src/pages/Calendar/Calendar.jsx
@@ -1,0 +1,9 @@
+import Navigation from '../../components/Common/Navbar'
+
+export default function Calendar(){
+  return (
+    <>
+      <Navigation />
+    </>
+  )
+}
