+++ src/pages/Login/IntroLoginPage.jsx
@@ -37,7 +37,7 @@ const IntroLoginPage = () => {
   useEffect(() => {
     if (userInfo && localStorage.getItem("token")) {
       // 로그인된 사용자인 경우 홈으로 이동
-      navigate("/home");
+      navigate("/calendar");
     }
   }, [userInfo, navigate]);
